requirements.txt
moviepy[optional]
git+https://github.com/openai/whisper.git
git+https://github.com/openai/CLIP.git
numpy
opencv-python
shutil
torch
transformers
tqdm
pillow
scikit-learn
traceback
shutil
re